backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0

# 数据库
PyMySQL>=1.1.0
cryptography>=41.0.0

# AI 集成
requests>=2.31.0
openai>=1.0.0

# 图片处理
Pillow>=10.0.0

# 工具
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0

# 开发测试
pytest>=7.4.0
pytest-django>=4.7.0
